# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/universities": {
      "get": {
        "operationId": "UniversitiesController_getUniversities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of universities retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all universities",
        "tags": [
          "Universities"
        ]
      },
      "post": {
        "operationId": "UniversitiesController_createUniversity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUniversityRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "University created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a new university (admin only)",
        "tags": [
          "Universities"
        ]
      }
    },
    "/api/universities/{id}": {
      "get": {
        "operationId": "UniversitiesController_getUniversityById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "University ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "University retrieved successfully"
          },
          "404": {
            "description": "University not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get university by ID",
        "tags": [
          "Universities"
        ]
      },
      "put": {
        "operationId": "UniversitiesController_updateUniversity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "University ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUniversityRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "University updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "University not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Update an existing university (admin only)",
        "tags": [
          "Universities"
        ]
      },
      "delete": {
        "operationId": "UniversitiesController_deleteUniversity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "University ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "University and linked programs deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "University not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a university and all linked programs (admin only)",
        "tags": [
          "Universities"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your Supabase access token"
      }
    },
    "schemas": {
...
      "CreateUniversityRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "University name"
          },
          "university_type": {
            "type": "string",
            "description": "University type",
            "enum": [
              "public",
              "private"
            ]
          },
          "based_in": {
            "type": "object",
            "description": "Based in (country/region)"
          },
          "state": {
            "type": "object",
            "description": "State"
          },
          "city": {
            "type": "object",
            "description": "City"
          },
          "address": {
            "type": "object",
            "description": "Full address"
          },
          "description": {
            "type": "object",
            "description": "University description"
          },
          "website_url": {
            "type": "object",
            "description": "Website URL"
          },
          "email": {
            "type": "object",
            "description": "Contact email"
          },
          "phone_number": {
            "type": "object",
            "description": "Contact phone number"
          },
          "average_fee": {
            "type": "number",
            "description": "Average fee (RM)"
          },
          "logo_url": {
            "type": "object",
            "description": "Logo URL"
          },
          "image_urls": {
            "type": "object",
            "description": "Image URLs (JSON array)"
          }
        },
        "required": [
          "name",
          "university_type"
        ]
      },
      "UpdateUniversityRequestDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}